# apps/api/app/routes/flows.py
"""POST /flows, GET /flows/{id}, PATCH /flows/{id}."""

from __future__ import annotations

import logging
from datetime import UTC, datetime

from fastapi import APIRouter, Depends, HTTPException, Request
from py2dataiku.exceptions import Py2DataikuError
from py2dataiku.models.dataiku_flow import DataikuFlow

from ..deps import get_audit_repo, get_flows_repo
from ..schemas.flows import (
    CreatedFlowResponse,
    SavedFlowResponse,
    SaveFlowRequest,
    UpdateFlowRequest,
)
from ..store import AuditEvent, AuditRepo, FlowsRepo

logger = logging.getLogger(__name__)

router = APIRouter(prefix="/flows", tags=["flows"])


def _client_actor(request: Request) -> str:
    """Best-effort actor identity for audit purposes."""
    return request.headers.get("X-Actor") or (
        request.client.host if request.client else "anonymous"
    )


def _hydrate(flow_dict: dict) -> dict:
    """Round-trip *flow_dict* through ``DataikuFlow.from_dict`` to validate.

    Raises ``HTTPException(422)`` on a malformed flow body.
    """
    try:
        flow = DataikuFlow.from_dict(flow_dict)
    except (Py2DataikuError, ValueError, KeyError, TypeError) as exc:
        raise HTTPException(
            status_code=422, detail=f"Malformed flow body: {exc}"
        ) from exc
    return flow.to_dict()


@router.post(
    "",
    response_model=CreatedFlowResponse,
    status_code=201,
    summary="Persist a flow",
)
def post_flow(
    body: SaveFlowRequest,
    request: Request,
    flows: FlowsRepo = Depends(get_flows_repo),
    audit: AuditRepo = Depends(get_audit_repo),
) -> CreatedFlowResponse:
    flow_dict = body.flow.model_dump(by_alias=True)
    _hydrate(flow_dict)  # validate, but persist the raw dict
    record = flows.save(flow=flow_dict, name=body.name, tags=list(body.tags or []))
    audit.append(
        AuditEvent(
            actor=_client_actor(request),
            action="flow.create",
            resource_type="flow",
            resource_id=record.id,
            details={"name": record.name, "tags": list(record.tags)},
        )
    )
    return CreatedFlowResponse(id=record.id, created_at=record.created_at)


@router.get("/{flow_id}", response_model=SavedFlowResponse, summary="Read a saved flow")
def get_flow(
    flow_id: str,
    flows: FlowsRepo = Depends(get_flows_repo),
) -> SavedFlowResponse:
    record = flows.get(flow_id)
    if record is None:
        raise HTTPException(status_code=404, detail=f"flow '{flow_id}' not found")
    # Round-trip through DataikuFlow to keep parity with /convert responses.
    hydrated = _hydrate(record.flow)
    return SavedFlowResponse(
        id=record.id,
        name=record.name,
        flow=hydrated,  # type: ignore[arg-type]
        created_at=record.created_at,
        updated_at=record.updated_at,
        tags=list(record.tags),
    )


@router.patch(
    "/{flow_id}", response_model=SavedFlowResponse, summary="Patch a saved flow"
)
def patch_flow(
    flow_id: str,
    body: UpdateFlowRequest,
    request: Request,
    flows: FlowsRepo = Depends(get_flows_repo),
    audit: AuditRepo = Depends(get_audit_repo),
) -> SavedFlowResponse:
    if body.flow is None and body.name is None and body.tags is None:
        raise HTTPException(
            status_code=422, detail="At least one of flow, name, tags must be set"
        )
    new_flow_dict: dict | None = None
    if body.flow is not None:
        candidate = body.flow.model_dump(by_alias=True)
        new_flow_dict = _hydrate(candidate)
    try:
        record = flows.update(
            flow_id,
            flow=new_flow_dict,
            name=body.name,
            tags=body.tags,
        )
    except KeyError as exc:
        raise HTTPException(
            status_code=404, detail=f"flow '{flow_id}' not found"
        ) from exc
    audit.append(
        AuditEvent(
            actor=_client_actor(request),
            action="flow.update",
            resource_type="flow",
            resource_id=record.id,
            details={
                "name": record.name,
                "tags": list(record.tags),
                "fields": [
                    k
                    for k, v in {
                        "flow": body.flow,
                        "name": body.name,
                        "tags": body.tags,
                    }.items()
                    if v is not None
                ],
            },
        )
    )
    return SavedFlowResponse(
        id=record.id,
        name=record.name,
        flow=record.flow,  # type: ignore[arg-type]
        created_at=record.created_at,
        updated_at=record.updated_at,
        tags=list(record.tags),
    )