# packages/types/openapi.snapshot.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "py-iku Studio API",
    "description": "FastAPI wrapper around py2dataiku for py-iku Studio",
    "version": "0.0.0"
  },
  "paths": {
...
    "/flows": {
      "post": {
        "tags": [
          "flows"
        ],
        "summary": "Persist a flow",
        "operationId": "post_flow_flows_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveFlowRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedFlowResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/flows/{flow_id}": {
      "get": {
        "tags": [
          "flows"
        ],
        "summary": "Read a saved flow",
        "operationId": "get_flow_flows__flow_id__get",
        "parameters": [
          {
            "name": "flow_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Flow Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedFlowResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "flows"
        ],
        "summary": "Patch a saved flow",
        "operationId": "patch_flow_flows__flow_id__patch",
        "parameters": [
          {
            "name": "flow_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Flow Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFlowRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedFlowResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ColumnSchemaModel": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "nullable": {
            "type": "boolean",
            "title": "Nullable",
            "default": true
          },
          "default": {
            "title": "Default"
          },
          "format": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Format"
          }
        },
        "type": "object",
        "required": [
          "name",
          "type"
        ],
        "title": "ColumnSchemaModel",
        "description": "Mirrors ColumnSchema.to_dict()."
      },
...
      "CreatedFlowResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "created_at"
        ],
        "title": "CreatedFlowResponse",
        "description": "Slim response from ``POST /flows``."
      },
      "DataikuDatasetModel": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "type": {
            "$ref": "#/components/schemas/DatasetTypeEnum"
          },
          "connection_type": {
            "$ref": "#/components/schemas/DatasetConnectionTypeEnum",
            "default": "Filesystem"
          },
          "schema": {
            "items": {
              "$ref": "#/components/schemas/ColumnSchemaModel"
            },
            "type": "array",
            "title": "Schema"
          },
          "source_variable": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Variable"
          },
          "source_line": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Line"
          },
          "notes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "name",
          "type"
        ],
        "title": "DataikuDatasetModel",
        "description": "Mirrors DataikuDataset.to_dict() exactly."
      },
      "DataikuFlowModel-Input": {
        "properties": {
          "flow_name": {
            "type": "string",
            "title": "Flow Name",
            "default": "converted_flow"
          },
          "generated_from": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generated From"
          },
          "generation_timestamp": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generation Timestamp"
          },
          "total_recipes": {
            "type": "integer",
            "title": "Total Recipes",
            "default": 0
          },
          "total_datasets": {
            "type": "integer",
            "title": "Total Datasets",
            "default": 0
          },
          "datasets": {
            "items": {
              "$ref": "#/components/schemas/DataikuDatasetModel"
            },
            "type": "array",
            "title": "Datasets"
          },
          "recipes": {
            "items": {
              "$ref": "#/components/schemas/DataikuRecipeModel-Input"
            },
            "type": "array",
            "title": "Recipes"
          },
          "optimization_notes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Optimization Notes"
          },
          "recommendations": {
            "items": {
              "$ref": "#/components/schemas/FlowRecommendationModel"
            },
            "type": "array",
            "title": "Recommendations"
          },
          "zones": {
            "items": {
              "$ref": "#/components/schemas/FlowZoneModel"
            },
            "type": "array",
            "title": "Zones"
          }
        },
        "type": "object",
        "title": "DataikuFlowModel",
        "description": "Mirrors DataikuFlow.to_dict() shape exactly.\n\nValidates that all recipe inputs/outputs reference datasets that exist\nin the datasets list."
      },
      "DataikuFlowModel-Output": {
        "properties": {
          "flow_name": {
            "type": "string",
            "title": "Flow Name",
            "default": "converted_flow"
          },
          "generated_from": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generated From"
          },
          "generation_timestamp": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generation Timestamp"
          },
          "total_recipes": {
            "type": "integer",
            "title": "Total Recipes",
            "default": 0
          },
          "total_datasets": {
            "type": "integer",
            "title": "Total Datasets",
            "default": 0
          },
          "datasets": {
            "items": {
              "$ref": "#/components/schemas/DataikuDatasetModel"
            },
            "type": "array",
            "title": "Datasets"
          },
          "recipes": {
            "items": {
              "$ref": "#/components/schemas/DataikuRecipeModel-Output"
            },
            "type": "array",
            "title": "Recipes"
          },
          "optimization_notes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Optimization Notes"
          },
          "recommendations": {
            "items": {
              "$ref": "#/components/schemas/FlowRecommendationModel"
            },
            "type": "array",
            "title": "Recommendations"
          },
          "zones": {
            "items": {
              "$ref": "#/components/schemas/FlowZoneModel"
            },
            "type": "array",
            "title": "Zones"
          }
        },
        "type": "object",
        "title": "DataikuFlowModel",
        "description": "Mirrors DataikuFlow.to_dict() shape exactly.\n\nValidates that all recipe inputs/outputs reference datasets that exist\nin the datasets list."
      },
      "DataikuRecipeModel-Input": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "type": {
            "$ref": "#/components/schemas/RecipeTypeEnum"
          },
          "inputs": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Inputs"
          },
          "outputs": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Outputs"
          },
          "source_lines": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Source Lines"
          },
          "notes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Notes"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1,
                "minimum": 0
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence",
            "description": "LLM self-reported mapping confidence in [0.0, 1.0]. None means rule-based or unspecified. UI shading bands: >=0.85 high (no shade), 0.60-0.84 medium (warn), <0.60 low (danger)."
          },
          "reasoning": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reasoning",
            "description": "One-sentence rationale for the recipe mapping. None for rule-based recipes."
          },
          "settings": {
            "anyOf": [
              {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/PrepareSettingsModel"
                  },
                  {
                    "$ref": "#/components/schemas/GroupingSettingsModel"
                  },
                  {
                    "$ref": "#/components/schemas/JoinSettingsModel"
                  },
                  {
                    "$ref": "#/components/schemas/WindowSettingsModel"
                  },
                  {
                    "$ref": "#/components/schemas/SamplingSettingsModel"
                  },
                  {
                    "$ref": "#/components/schemas/SplitSettingsModel"
                  },
                  {
                    "$ref": "#/components/schemas/SortSettingsModel"
                  },
                  {
                    "$ref": "#/components/schemas/TopNSettingsModel"
                  },
                  {
                    "$ref": "#/components/schemas/DistinctSettingsModel"
                  },
                  {
                    "$ref": "#/components/schemas/StackSettingsModel"
                  },
                  {
                    "$ref": "#/components/schemas/PythonSettingsModel"
                  },
                  {
                    "$ref": "#/components/schemas/PivotSettingsModel"
                  }
                ],
                "discriminator": {
                  "propertyName": "kind",
                  "mapping": {
                    "distinct": "#/components/schemas/DistinctSettingsModel",
                    "grouping": "#/components/schemas/GroupingSettingsModel",
                    "join": "#/components/schemas/JoinSettingsModel",
                    "pivot": "#/components/schemas/PivotSettingsModel",
                    "prepare": "#/components/schemas/PrepareSettingsModel",
                    "python": "#/components/schemas/PythonSettingsModel",
                    "sampling": "#/components/schemas/SamplingSettingsModel",
                    "sort": "#/components/schemas/SortSettingsModel",
                    "split": "#/components/schemas/SplitSettingsModel",
                    "stack": "#/components/schemas/StackSettingsModel",
                    "topn": "#/components/schemas/TopNSettingsModel",
                    "window": "#/components/schemas/WindowSettingsModel"
                  }
                }
              },
              {
                "type": "null"
              }
            ],
            "title": "Settings",
            "description": "Composed recipe settings (discriminated union over 12 subclasses). Not populated by py-iku to_dict(); present here to expose the full settings schema contract in /openapi.json."
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "name",
          "type"
        ],
        "title": "DataikuRecipeModel",
        "description": "Mirrors DataikuRecipe.to_dict() shape.\n\npy-iku's to_dict() inlines recipe-type-specific settings via to_display_dict()\nrather than nesting them under a 'settings' key. Those fields pass through via\n`extra=\"allow\"`.\n\nThe `settings` field is declared here as Optional so all 12 RecipeSettings\nsubclass schemas are registered in the OpenAPI component registry. In practice\nit will be None for flows produced by py-iku's to_dict(), but callers that\nconstruct recipes directly may populate it."
      },
      "DataikuRecipeModel-Output": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "type": {
            "$ref": "#/components/schemas/RecipeTypeEnum"
          },
          "inputs": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Inputs"
          },
          "outputs": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Outputs"
          },
          "source_lines": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Source Lines"
          },
          "notes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Notes"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1,
                "minimum": 0
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence",
            "description": "LLM self-reported mapping confidence in [0.0, 1.0]. None means rule-based or unspecified. UI shading bands: >=0.85 high (no shade), 0.60-0.84 medium (warn), <0.60 low (danger)."
          },
          "reasoning": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reasoning",
            "description": "One-sentence rationale for the recipe mapping. None for rule-based recipes."
          },
          "settings": {
            "anyOf": [
              {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/PrepareSettingsModel"
                  },
                  {
                    "$ref": "#/components/schemas/GroupingSettingsModel"
                  },
                  {
                    "$ref": "#/components/schemas/JoinSettingsModel"
                  },
                  {
                    "$ref": "#/components/schemas/WindowSettingsModel"
                  },
                  {
                    "$ref": "#/components/schemas/SamplingSettingsModel"
                  },
                  {
                    "$ref": "#/components/schemas/SplitSettingsModel"
                  },
                  {
                    "$ref": "#/components/schemas/SortSettingsModel"
                  },
                  {
                    "$ref": "#/components/schemas/TopNSettingsModel"
                  },
                  {
                    "$ref": "#/components/schemas/DistinctSettingsModel"
                  },
                  {
                    "$ref": "#/components/schemas/StackSettingsModel"
                  },
                  {
                    "$ref": "#/components/schemas/PythonSettingsModel"
                  },
                  {
                    "$ref": "#/components/schemas/PivotSettingsModel"
                  }
                ],
                "discriminator": {
                  "propertyName": "kind",
                  "mapping": {
                    "distinct": "#/components/schemas/DistinctSettingsModel",
                    "grouping": "#/components/schemas/GroupingSettingsModel",
                    "join": "#/components/schemas/JoinSettingsModel",
                    "pivot": "#/components/schemas/PivotSettingsModel",
                    "prepare": "#/components/schemas/PrepareSettingsModel",
                    "python": "#/components/schemas/PythonSettingsModel",
                    "sampling": "#/components/schemas/SamplingSettingsModel",
                    "sort": "#/components/schemas/SortSettingsModel",
                    "split": "#/components/schemas/SplitSettingsModel",
                    "stack": "#/components/schemas/StackSettingsModel",
                    "topn": "#/components/schemas/TopNSettingsModel",
                    "window": "#/components/schemas/WindowSettingsModel"
                  }
                }
              },
              {
                "type": "null"
              }
            ],
            "title": "Settings",
            "description": "Composed recipe settings (discriminated union over 12 subclasses). Not populated by py-iku to_dict(); present here to expose the full settings schema contract in /openapi.json."
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "name",
          "type"
        ],
        "title": "DataikuRecipeModel",
        "description": "Mirrors DataikuRecipe.to_dict() shape.\n\npy-iku's to_dict() inlines recipe-type-specific settings via to_display_dict()\nrather than nesting them under a 'settings' key. Those fields pass through via\n`extra=\"allow\"`.\n\nThe `settings` field is declared here as Optional so all 12 RecipeSettings\nsubclass schemas are registered in the OpenAPI component registry. In practice\nit will be None for flows produced by py-iku's to_dict(), but callers that\nconstruct recipes directly may populate it."
      },
      "DatasetConnectionTypeEnum": {
        "type": "string",
        "enum": [
          "Filesystem",
          "PostgreSQL",
          "MySQL",
          "BigQuery",
          "Snowflake",
          "Redshift",
          "S3",
          "GCS",
          "Azure",
          "HDFS",
          "ManagedFolder",
          "MongoDB",
          "Elasticsearch"
        ],
        "title": "DatasetConnectionTypeEnum",
        "description": "Dataset connection types — mirrors DatasetConnectionType enum values."
      },
      "DatasetTypeEnum": {
        "type": "string",
        "enum": [
          "input",
          "intermediate",
          "output"
        ],
        "title": "DatasetTypeEnum",
        "description": "Dataset types — mirrors DatasetType enum values."
      },
...
      "FlowRecommendationModel": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type"
          },
          "priority": {
            "type": "string",
            "title": "Priority"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "impact": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Impact"
          },
          "action": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Action"
          },
          "source_lines": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Source Lines"
          }
        },
        "type": "object",
        "required": [
          "type",
          "priority",
          "message"
        ],
        "title": "FlowRecommendationModel",
        "description": "Mirrors FlowRecommendation.to_dict()."
      },
      "FlowZoneModel": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "color": {
            "type": "string",
            "title": "Color",
            "default": "#4b96e6"
          },
          "datasets": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Datasets"
          },
          "recipes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Recipes"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "FlowZoneModel",
        "description": "Mirrors FlowZone.to_dict()."
      },
      "GroupingSettingsModel": {
        "properties": {
          "kind": {
            "type": "string",
            "const": "grouping",
            "title": "Kind",
            "default": "grouping"
          },
          "keys": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Keys"
          },
          "aggregations": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Aggregations"
          }
        },
        "type": "object",
        "title": "GroupingSettingsModel",
        "description": "Mirrors GroupingSettings.to_display_dict(): keys + aggregations."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PrepareSettingsModel": {
        "properties": {
          "kind": {
            "type": "string",
            "const": "prepare",
            "title": "Kind",
            "default": "prepare"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/PrepareStepModel"
            },
            "type": "array",
            "title": "Steps"
          },
          "step_count": {
            "type": "integer",
            "title": "Step Count",
            "default": 0
          }
        },
        "type": "object",
        "title": "PrepareSettingsModel",
        "description": "Mirrors PrepareSettings.to_display_dict(): steps + step_count."
      },
...
      "RecipeTypeEnum": {
        "type": "string",
        "enum": [
          "prepare",
          "sync",
          "grouping",
          "window",
          "join",
          "fuzzyjoin",
          "geojoin",
          "stack",
          "split",
          "sort",
          "distinct",
          "topn",
          "pivot",
          "sampling",
          "download",
          "generate_features",
          "generate_statistics",
          "push_to_editable",
          "list_folder_contents",
          "dynamic_repeat",
          "extract_failed_rows",
          "upsert",
          "list_access",
          "python",
          "r",
          "sql_script",
          "hive",
          "impala",
          "spark_sql_query",
          "pyspark",
          "spark_scala",
          "sparkr",
          "shell",
          "prediction_scoring",
          "clustering_scoring",
          "standalone_evaluation",
          "ai_assistant_generate"
        ],
        "title": "RecipeTypeEnum",
        "description": "All 37 Dataiku DSS recipe types."
      },
...
      "SaveFlowRequest": {
        "properties": {
          "flow": {
            "$ref": "#/components/schemas/DataikuFlowModel-Input"
          },
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          }
        },
        "type": "object",
        "required": [
          "flow",
          "name"
        ],
        "title": "SaveFlowRequest",
        "description": "Request body for ``POST /flows``."
      },
      "SavedFlowResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "flow": {
            "$ref": "#/components/schemas/DataikuFlowModel-Output"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "flow",
          "created_at",
          "updated_at"
        ],
        "title": "SavedFlowResponse",
        "description": "Full saved-flow record."
      },
...
      "SplitSettingsModel": {
        "properties": {
          "kind": {
            "type": "string",
            "const": "split",
            "title": "Kind",
            "default": "split"
          },
          "splitMode": {
            "type": "string",
            "title": "Splitmode",
            "default": "FILTER"
          },
          "condition": {
            "type": "string",
            "title": "Condition",
            "default": ""
          }
        },
        "type": "object",
        "title": "SplitSettingsModel",
        "description": "Mirrors SplitSettings.to_dict(): splitMode + condition."
      },
...
      "UpdateFlowRequest": {
        "properties": {
          "flow": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DataikuFlowModel-Input"
              },
              {
                "type": "null"
              }
            ]
          },
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "tags": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags"
          }
        },
        "type": "object",
        "title": "UpdateFlowRequest",
        "description": "Request body for ``PATCH /flows/{id}`` — every field is optional."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}